refactor: split view edit-event
